Sparse: Implement simple unary ufuncs operators
By "simple" I mean operators that map 0->0 so we can implement it by
just re-dispatching on the values tensor. That does mean we have `sin`
but not `cos` for example, but without fill value support this is the
best that can be done.

Most of these don't support autograd because the derivative formulas
use unsupported operators.

ghstack-source-id: bb20a99ae1702e5578ccf10971ac745e82bdf1e1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68887